fix: Fix wrong casting strategy when using avg aggregation on int fields. Pre-catch metric aggregations on text or keyword fields

+++ src/Common/Aggregation/AggregationCaster.php
@@ -17,7 +17,8 @@ class AggregationCaster
 
         switch ($type) {
             case FieldType::INT:
-                $caster = fn($value) => (int) $value;
+                // cast to float, because of avg aggregation
+                $caster = fn($value) => (float) $value;
                 break;
             case FieldType::FLOAT:
                 $caster = fn($value) => round((float) $value, 6);
